--- a/packages/core/chrome/core-chrome-browser-internal/src/project_navigation/utils.ts
+++ b/packages/core/chrome/core-chrome-browser-internal/src/project_navigation/utils.ts
@@ -114,7 +114,6 @@ export const findActiveNodes = (
 
   const activeNodeFromKey = (key: string): ChromeProjectNavigationNode => ({
     ...navTree[key],
-    isActive: true,
   });
 
   Object.entries(navTree).forEach(([key, node]) => {