--- a/packages/shared-ux/chrome/navigation/mocks/src/storybook.ts
+++ b/packages/shared-ux/chrome/navigation/mocks/src/storybook.ts
@@ -14,7 +14,7 @@ import { NavigationServices } from '../../types';
 type Arguments = NavigationServices;
 export type Params = Pick<
   Arguments,
-  'navIsOpen' | 'recentlyAccessed$' | 'activeNodes$' | 'navLinks$' | 'onProjectNavigationChange'
+  'navIsOpen' | 'recentlyAccessed$' | 'activeNodes$' | 'deepLinks$' | 'onProjectNavigationChange'
 >;
 
 export class StorybookMock extends AbstractStorybookMock<{}, NavigationServices> {
@@ -41,7 +41,7 @@ export class StorybookMock extends AbstractStorybookMock<{}, NavigationServices>
       basePath: { prepend: (suffix: string) => `/basepath${suffix}` },
       navigateToUrl,
       recentlyAccessed$: params.recentlyAccessed$ ?? new BehaviorSubject([]),
-      navLinks$: params.navLinks$ ?? new BehaviorSubject([]),
+      deepLinks$: params.deepLinks$ ?? new BehaviorSubject({}),
       onProjectNavigationChange: params.onProjectNavigationChange ?? (() => undefined),
       activeNodes$: params.activeNodes$ ?? new BehaviorSubject([]),
       isSideNavCollapsed: true,