--- a/packages/shared-ux/chrome/navigation/src/ui/navigation.stories.tsx
+++ b/packages/shared-ux/chrome/navigation/src/ui/navigation.stories.tsx
@@ -16,7 +16,6 @@ import {
   EuiButton,
   EuiCollapsibleNavBeta,
   EuiCollapsibleNavBetaProps,
-  EuiCollapsibleNavItem,
   EuiFlexGroup,
   EuiFlexItem,
   EuiHeader,