… packages/core/chrome/core-chrome-browser-internal/src/project_navigation/project_navigation_service.test.ts
… packages/core/chrome/core-chrome-browser-internal/src/project_navigation/utils.test.ts
… packages/shared-ux/chrome/navigation/__jest__/__snapshots__/project_navigation.test.tsx.snap
… packages/shared-ux/chrome/navigation/__jest__/active_node.test.tsx
… packages/shared-ux/chrome/navigation/__jest__/build_nav_tree.test.tsx
… packages/shared-ux/chrome/navigation/__jest__/links.test.tsx
… packages/shared-ux/chrome/navigation/__jest__/panel.test.tsx
… packages/shared-ux/chrome/navigation/__jest__/project_navigation.test.tsx
… packages/shared-ux/chrome/navigation/src/ui/components/navigation_item.tsx
… packages/shared-ux/chrome/navigation/src/ui/components/navigation_item_open_panel.tsx
… packages/shared-ux/chrome/navigation/src/ui/components/navigation_section_ui.tsx
… packages/shared-ux/chrome/navigation/src/ui/components/panel/context.tsx
… packages/shared-ux/chrome/navigation/src/ui/components/panel/default_content.tsx
… packages/shared-ux/chrome/navigation/src/ui/components/panel/navigation_panel.tsx
… packages/shared-ux/chrome/navigation/src/ui/default_navigation.tsx
… packages/shared-ux/chrome/navigation/src/ui/hooks/index.ts
… packages/shared-ux/chrome/navigation/src/ui/hooks/use_init_navnode.ts
… packages/shared-ux/chrome/navigation/src/ui/hooks/use_register_tree_node.ts
… packages/shared-ux/chrome/navigation/src/ui/types.ts
… packages/shared-ux/chrome/navigation/src/utils.ts
… packages/shared-ux/chrome/navigation/types/index.ts
… x-pack/plugins/security_solution_serverless/public/navigation/navigation_tree/chrome_navigation_tree.test.ts
… x-pack/plugins/security_solution_serverless/public/navigation/navigation_tree/chrome_navigation_tree.ts
… x-pack/plugins/security_solution_serverless/public/navigation/project_navigation/project_navigation.tsx
… x-pack/test_serverless/functional/services/ml/observability_navigation.ts