As stated in https://github.com/elastic/kibana/issues/171169, when working on the stateful side nav POC (https://github.com/elastic/kibana/pull/170925) it became clear that the performance of the side navigation for the new "project" type was not optimal.

In this PR I've revisited the architecture on how the navigation nodes are registered and removed all possible asynchronicity to render the nodes in the UI. We used to have for each and every node declared in the navigation tree definition (independently if they are rendered or not in the UI) 3 states (`path`, `isActive`, `childrenNodes`) + an asynchronous mechanism to register itself to its parent through `useEffect()`). All that was far from performant and the problem grew for each node added.

The refactor that I made also greatly improves the implementation of the side navigation.

## Benchmark

This is the result of the benchmark before and after applying the changes in this PR to the POC. The time calculated is the time it takes to switch to that navigation.

### Before

Search: 524ms (switch from "Classic" nav)
Obs: 680ms
Security: 1410ms

Search: 397ms
Obs: 758ms
Security: 1560ms

Average:
Search: 460ms
Obs: 719ms
Security: 1485ms


### After

Search: 358ms (switch from "Classic" nav)
Obs: 269ms
Security: 240ms

Search: 264ms
Obs: 328ms
Security: 346ms

Average:
Search: 311ms (**148%** faster 🚀 )
Obs: 298ms (**241%** faster 🚀 )
Security: 293ms (**506%** faster 🚀 )

## Video

This is how it "felt"

### Before

https://github.com/elastic/kibana/assets/2854616/5000ffe5-a92c-4460-9225-1ca430afc016


### After


https://github.com/elastic/kibana/assets/2854616/2663c906-dca8-4704-9290-cad9848a12fc

## Performance

### Before

<img width="2556" alt="Screenshot 2023-11-24 at 10 34 09" src="https://github.com/elastic/kibana/assets/2854616/3f89dc89-91c3-459c-9bcf-c7d1a4ba1726">

### After

<img width="2558" alt="Screenshot 2023-11-24 at 10 29 16" src="https://github.com/elastic/kibana/assets/2854616/ec0a4556-1f3a-4578-b715-22a74d0c58cb">

Fixes https://github.com/elastic/kibana/issues/171169