@sebelga 2023-11-27
Thanks for the review @Dosant @vadimkibana @eokoneyo and @semd !

Anton, regarding your comment on the breadcrumb issue I'll look at it in a separate PR. 
At first sight on your screenshot, it almost look that it is a good thing to not have a breadcrumb in a smaller screen as there isn't much space anyway for it. Let me see what we can do here.